// src/types/event.ts
export interface IEvent {
  no_event: string;
  no_team_sort: string;
  event_name: string;
  no_match: string;
  match_date: string;
  club_home: string;
  club_away: string;
  home_score: string;
  away_score: string;
  ev_round: string;
  handicap_display: string;
  odds_home: string;
  odds_away: string;
  price_alert_flag_home: string;
  price_alert_flag_away: string;
}